pyproject.toml
[tool.poetry]
name = "pomodoro-analytics"
version = "0.1.0"
description = ""
authors = ["a.kapral26 <a.kapral26@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.4" }
pydantic-settings = "^2.6.1"
watchfiles = "^0.24.0"
ruff = "^0.7.3"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
faker = "^33.0.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Включить правила pycodestyle (E) и Pyflakes (F) по умолчанию.
# В отличие от Flake8, Ruff не включает предупреждения pycodestyle (W) или
# McCabe сложности (C901) по умолчанию.
# https://docs.astral.sh/ruff/settings/#tab-size
lint.select = [
    # pycodestyle
    "W",
    # Pyflakes
    "B",
    # flake8-bugbear
    "Q",
    # flake8-quotes
    "N",
    # pep8-naming
    "D",
    # pydocstyle
    "UP",
    # pyupgrade
    "S",
    # pyupgrade
    "T20",
    "RUF",
    # Ruff-specific rules
    "FLY",
    # flynt
    "TRY",
    # tryceratops
    "PD",
    # pandas-vet
    "TCH",
    # flake8-type-checking
    "ANN", # flake8-annotations
]

[lint]
# Разрешить автоматическое исправление всех включенных правил (при использовании `--fix`)
fixable = ["ALL"]
# Игнорировать следующие правила
ignore = [
    "F",
    "D107",
    "S608",
    "D212",
    "D205",
    "D400",
    "B904",
    "D415",
    "E722",
    "F401",
    "RUF001",
    "RUF002",
    "RUF003",
    "ANN101",
    "ANN002",
    "ANN003",
    "ANN204",
    "TRY003",
    "TRY200",
    "B904",
]
# Исключить следующие директории
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = { }
# Игнорировать следующие файлы
unfixable = []

[format]
# Включить переформатирование кода в docstring
docstring-code-format = true
# Использовать табуляцию вместо 4 пробелов для отступов
indent-style = "tab"
# Стиль кавычек
quote-style = "double"
# Разрешить неиспользуемые переменные с префиксом "_"
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Переносить аргументы методов/функций на новую строку
max-function-args = 1
max-function-arg-length = 40